+++ CHANGES.rst
@@ -4,8 +4,8 @@ Products.MeetingLiege Changelog
 4.2.16.2 (unreleased)
 ---------------------
 
-- Nothing changed yet.
-
+- Pre-configure `labels` for `Council` in `Migrate_To_4203._configureLabels`.
+  [gbastien]
 
 4.2.16.1 (2025-12-16)
 ---------------------